agentic-shared(test-credentials-directly,non_mechanical_provable): test the credential contract where it lives
The contract tests move from the OpenEnv example's test package (where they
reached the module through the launcher's re-exports) to
tests/fast/rollout/agentic/test_credentials.py, importing it directly, and
grow the coverage the module lacked: preflight_sdk on both outcomes,
resolve_provider_api_key (env wins / file fallback / default-path expansion /
missing names both supplies), and an unreadable key path counting as absent.
The example keeps only its own obligations: every registered backend has a
spec, and every spec names a real launcher arg.
